backend/requirements.txt
biopython>=1.83
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
matplotlib>=3.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
ruff>=0.2.0
black>=24.1.0
